Refactor partial file search into PartialTemplatesFinder
Also add some tests to assert that the cache is actually being used
rather than re-reading the partial template files
